--- src/main/java/wooteco/subway/domain/PathFinder.java
@@ -1,7 +1,0 @@
-package wooteco.subway.domain;
-
-import java.util.List;
-
-public interface PathFinder {
-    Path findShortest(List<Line> lines, Station source, Station target);
-}
