+++ src/main/java/wooteco/subway/domain/farerule/AgeDiscount.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain;
+package wooteco.subway.domain.farerule;
 
 import java.util.Arrays;
 import java.util.function.Function;
