+++ src/main/java/wooteco/subway/domain/pathfinder/PathFinder.java
@@ -1,0 +1,10 @@
+package wooteco.subway.domain.pathfinder;
+
+import java.util.List;
+import wooteco.subway.domain.Line;
+import wooteco.subway.domain.Path;
+import wooteco.subway.domain.Station;
+
+public interface PathFinder {
+    Path findShortest(List<Line> lines, Station source, Station target);
+}
