+++ src/main/java/wooteco/subway/service/PathService.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import wooteco.subway.domain.Line;
 import wooteco.subway.domain.Path;
-import wooteco.subway.domain.PathFinder;
+import wooteco.subway.domain.pathfinder.PathFinder;
 import wooteco.subway.domain.Station;
 import wooteco.subway.dto.PathResponse;
 import wooteco.subway.repository.LineRepository;
